# data/processed/application/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MPLADS Sentinel API",
    "description": "Governed decision-support API over frozen MPLADS prototype intelligence artifacts.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/works/{work_id}/reviews": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "List Reviews",
        "operationId": "list_reviews_api_v1_works__work_id__reviews_get",
        "parameters": [
          {
            "name": "work_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Work Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewResponse"
                  },
                  "title": "Response List Reviews Api V1 Works  Work Id  Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Create Review",
        "operationId": "create_review_api_v1_works__work_id__reviews_post",
        "parameters": [
          {
            "name": "work_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Work Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewCreate": {
        "properties": {
          "actor_role": {
            "$ref": "#/components/schemas/Role"
          },
          "actor_label": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Actor Label"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "note": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 2,
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "actor_role",
          "actor_label",
          "status",
          "note"
        ],
        "title": "ReviewCreate"
      },
      "ReviewResponse": {
        "properties": {
          "review_id": {
            "type": "string",
            "title": "Review Id"
          },
          "work_id": {
            "type": "string",
            "title": "Work Id"
          },
          "actor_role": {
            "$ref": "#/components/schemas/Role"
          },
          "actor_label": {
            "type": "string",
            "title": "Actor Label"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "review_id",
          "work_id",
          "actor_role",
          "actor_label",
          "status",
          "note",
          "created_at"
        ],
        "title": "ReviewResponse"
      },
      "ReviewStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "IN_REVIEW",
          "NEEDS_CLARIFICATION",
          "ESCALATED",
          "RESOLVED",
          "FALSE_POSITIVE"
        ],
        "title": "ReviewStatus"
      },
      "Role": {
        "type": "string",
        "enum": [
          "MOSPI",
          "STATE",
          "DISTRICT",
          "MP"
        ],
        "title": "Role"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}